Allow client basic auth to be used in password grant flow
