+++ support/cas-server-support-oauth-core-api/src/main/java/org/apereo/cas/support/oauth/authenticator/OAuth20UsernamePasswordAuthenticator.java
@@ -13,11 +13,9 @@ import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import lombok.val;
 import org.apache.commons.lang3.StringUtils;
-import org.apache.commons.lang3.tuple.Pair;
 import org.pac4j.core.context.WebContext;
 import org.pac4j.core.credentials.UsernamePasswordCredentials;
 import org.pac4j.core.credentials.authenticator.Authenticator;
-import org.pac4j.core.credentials.extractor.BasicAuthExtractor;
 import org.pac4j.core.exception.CredentialsException;
 import org.pac4j.core.profile.CommonProfile;
 
@@ -42,7 +40,7 @@ public class OAuth20UsernamePasswordAuthenticator implements Authenticator<Usern
     public void validate(final UsernamePasswordCredentials credentials, final WebContext context) throws CredentialsException {
         val casCredential = new UsernamePasswordCredential(credentials.getUsername(), credentials.getPassword());
         try {
-            val clientIdAndSecret = getClientIdAndClientSecret(context);
+            val clientIdAndSecret = OAuth20Utils.getClientIdAndClientSecret(context);
             if (clientIdAndSecret == null || StringUtils.isBlank(clientIdAndSecret.getKey())) {
                 throw new CredentialsException("No client credentials could be identified in this request");
             }
@@ -81,25 +79,5 @@ public class OAuth20UsernamePasswordAuthenticator implements Authenticator<Usern
         } catch (final Exception e) {
             throw new CredentialsException("Cannot login user using CAS internal authentication", e);
         }
-    }
-
-    /**
-     * Gets client id and client secret.
-     *
-     * @param context the context
-     * @return the client id and client secret
-     */
-    protected Pair<String, String> getClientIdAndClientSecret(final WebContext context) {
-        val extractor = new BasicAuthExtractor();
-        val upcResult = extractor.extract(context);
-        if (upcResult.isPresent()) {
-            val upc = upcResult.get();
-            return Pair.of(upc.getUsername(), upc.getPassword());
-        }
-        val clientId = context.getRequestParameter(OAuth20Constants.CLIENT_ID)
-            .map(String::valueOf).orElse(StringUtils.EMPTY);
-        val clientSecret = context.getRequestParameter(OAuth20Constants.CLIENT_SECRET)
-            .map(String::valueOf).orElse(StringUtils.EMPTY);
-        return Pair.of(clientId, clientSecret);
     }
 }
